+++ site/examples/animation/morph/demo/morph.js
@@ -67,7 +67,7 @@ canvas.addEventListener(CanvasEvent.READY, () => {
   const path2Str = convertToPath(path2);
   const pathA = new Path({
     style: {
-      path: path1Str,
+      d: path1Str,
       stroke: '#F04864',
     },
   });
@@ -94,12 +94,12 @@ canvas.addEventListener(CanvasEvent.READY, () => {
   const linePathStr = convertToPath(line);
   const pathB = new Path({
     style: {
-      path: path1Str,
+      d: path1Str,
       stroke: '#F04864',
     },
   });
   canvas.appendChild(pathB);
-  pathB.animate([{ path: path1Str }, { path: linePathStr }], {
+  pathB.animate([{ d: path1Str }, { d: linePathStr }], {
     duration: 2500,
     easing: 'ease',
     iterations: Infinity,
@@ -120,7 +120,7 @@ canvas.addEventListener(CanvasEvent.READY, () => {
   const circlePathStr = convertToPath(circle);
   const pathC = new Path({
     style: {
-      path: path1Str,
+      d: path1Str,
       stroke: '#F04864',
     },
   });
@@ -152,12 +152,12 @@ canvas.addEventListener(CanvasEvent.READY, () => {
   );
   const pathD = new Path({
     style: {
-      path: circlePathStr,
+      d: circlePathStr,
       stroke: '#F04864',
     },
   });
   canvas.appendChild(pathD);
-  pathD.animate([{ path: circlePathStr }, { path: polylinePathStr }], {
+  pathD.animate([{ d: circlePathStr }, { d: polylinePathStr }], {
     duration: 2500,
     easing: 'ease',
     iterations: Infinity,
@@ -181,12 +181,12 @@ canvas.addEventListener(CanvasEvent.READY, () => {
   );
   const pathE = new Path({
     style: {
-      path: path1Str,
+      d: path1Str,
       stroke: '#F04864',
     },
   });
   canvas.appendChild(pathE);
-  pathE.animate([{ path: path1Str }, { path: polygonPathStr }], {
+  pathE.animate([{ d: path1Str }, { d: polygonPathStr }], {
     duration: 2500,
     easing: 'ease',
     iterations: Infinity,
@@ -209,7 +209,7 @@ canvas.addEventListener(CanvasEvent.READY, () => {
   );
   const pathF = new Path({
     style: {
-      path: rectPathStr,
+      d: rectPathStr,
       stroke: '#F04864',
       fill: '',
       opacity: 0.5,
@@ -219,8 +219,8 @@ canvas.addEventListener(CanvasEvent.READY, () => {
   canvas.appendChild(pathF);
   pathF.animate(
     [
-      { path: rectPathStr, stroke: '#F04864', fill: 'blue' },
-      { path: circlePathStr, stroke: 'blue', fill: '#F04864' },
+      { d: rectPathStr, stroke: '#F04864', fill: 'blue' },
+      { d: circlePathStr, stroke: 'blue', fill: '#F04864' },
     ],
     {
       duration: 2500,
@@ -235,22 +235,22 @@ canvas.addEventListener(CanvasEvent.READY, () => {
    */
   const starPath = new Path({
     style: {
-      path: 'M301.113,12.011l99.25,179.996l201.864,38.778L461.706,380.808l25.508,203.958l-186.101-87.287L115.01,584.766l25.507-203.958L0,230.785l201.86-38.778L301.113,12.011',
+      d: 'M301.113,12.011l99.25,179.996l201.864,38.778L461.706,380.808l25.508,203.958l-186.101-87.287L115.01,584.766l25.507-203.958L0,230.785l201.86-38.778L301.113,12.011',
     },
   });
   starPath.scale(0.2);
   starPath.translate(200, 200);
   const pathG = new Path({
     style: {
-      path: rectPathStr,
+      d: rectPathStr,
       lineWidth: 2,
     },
   });
   canvas.appendChild(pathG);
   pathG.animate(
     [
-      { path: rectPathStr, stroke: '#F04864', fill: 'blue' },
-      { path: convertToPath(starPath), stroke: 'blue', fill: '#F04864' },
+      { d: rectPathStr, stroke: '#F04864', fill: 'blue' },
+      { d: convertToPath(starPath), stroke: 'blue', fill: '#F04864' },
     ],
     {
       duration: 2500,