+++ site/examples/shape/ellipse/demo/ellipse.js
@@ -212,13 +212,11 @@ ellipseFolder
 
 const transformFolder = gui.addFolder('transform');
 const transformConfig = {
-  localPositionX: 300,
-  localPositionY: 200,
+  localPositionX: 0,
+  localPositionY: 0,
   localScale: 1,
   localEulerAngles: 0,
-  transformOrigin: 'center',
-  anchorX: 0.5,
-  anchorY: 0.5,
+  transformOrigin: 'left top',
 };
 transformFolder
   .add(transformConfig, 'transformOrigin', [
@@ -253,10 +251,4 @@ transformFolder
   .onChange((localEulerAngles) => {
     ellipse.setLocalEulerAngles(localEulerAngles);
   });
-transformFolder.add(transformConfig, 'anchorX', 0, 1).onChange((anchorX) => {
-  ellipse.style.anchor = [anchorX, transformConfig.anchorY];
-});
-transformFolder.add(transformConfig, 'anchorY', 0, 1).onChange((anchorY) => {
-  ellipse.style.anchor = [transformConfig.anchorX, anchorY];
-});
 transformFolder.open();