+++ packages/g-canvas/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @antv/g-canvas
 
+## 2.0.7
+
+### Patch Changes
+
+-   Updated dependencies [756ba5d3]
+    -   @antv/g-plugin-html-renderer@2.0.6
+
 ## 2.0.6
 
 ### Patch Changes