+++ packages/g-plugin-canvaskit-renderer/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @antv/g-plugin-canvaskit-renderer
 
+## 2.0.6
+
+### Patch Changes
+
+-   756ba5d3: Update deps.
+
 ## 2.0.5
 
 ### Patch Changes